requirements.txt
python-dotenv==1.0.0
slack-sdk==3.21.3
requests==2.31.0